requirements.txt
openai
googlesearch-python
python-dotenv
tweepy